tests/acceptance: allow forcing the use of s3 and overriding s3 server address
Adds new command line options:

--use-s3: use of s3 instead of starting own, local, HTTP server
--s3-address: allow changing of S3 address

This enables use of minio instead of s3 for transferring images under test.
There's still a small caveat, you still need to set bucket policy in minio to
something like Read-Only or Read-Write.

Changelog: none

+++ tests/acceptance/test_update.py
@@ -185,6 +185,7 @@ class TestUpdates:
     @pytest.mark.min_mender_version('1.0.0')
     def test_network_based_image_update(self, successful_image_update_mender, bitbake_variables):
         http_server_location = pytest.config.getoption("--http-server")
+        use_s3 = pytest.config.getoption("--use-s3")
         bbb = pytest.config.getoption("--bbb")
 
         if not env.host_string:
@@ -194,9 +195,10 @@ class TestUpdates:
 
         (active_before, passive_before) = determine_active_passive_part(bitbake_variables)
 
-        if bbb:
+        if bbb or use_s3:
             Helpers.upload_to_s3()
-            http_server_location = "s3.amazonaws.com/mender/temp"
+            s3_address = pytest.config.getoption("--s3-address")
+            http_server_location = "{}/mender/temp".format(s3_address)
         else:
             http_server = subprocess.Popen(["python", "-m", "SimpleHTTPServer"])
             assert(http_server)
@@ -204,7 +206,7 @@ class TestUpdates:
         try:
             run("mender -rootfs http://%s/successful_image_update.mender" % (http_server_location))
         finally:
-            if not bbb:
+            if not bbb and not use_s3:
                 http_server.terminate()
 
         output = run("fw_printenv bootcount")
